refactor: moved utils outside and renamed it to shared
